backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
pandas
scikit-learn
google-generativeai
python-dotenv